[IN PROGRESS] New Resource base, some other changes, struct additions, dead part refactoring

+++ core/src/main/java/com/draniksoft/ome/editor/support/track/UsageTracker.java
@@ -2,8 +2,8 @@ package com.draniksoft.ome.editor.support.track;
 
 public interface UsageTracker {
 
-    int usages();
+    short usages();
 
-    int usage(int delta);
+    short usage(short delta);
 
 }
